Add initial test of hypothesis module
